replace Stringify to models
